Move ContainerTemperatureDamageThresholdsComponent to Content.Shared
