+++ Content.Server/Temperature/Systems/TemperatureSystem.cs
@@ -1,6 +1,5 @@
 using Content.Server.Atmos.EntitySystems;
 using Content.Server.Body.Components;
-using Content.Server.Temperature.Components;
 using Content.Shared.Alert;
 using Content.Shared.Inventory;
 using Content.Shared.Temperature;
@@ -26,13 +25,6 @@ public sealed class TemperatureSystem : SharedTemperatureSystem
         SubscribeLocalEvent<AlertsComponent, OnTemperatureChangeEvent>(ServerAlert);
         SubscribeLocalEvent<TemperatureProtectionComponent, InventoryRelayedEvent<ModifyChangedTemperatureEvent>>(
             OnTemperatureChangeAttempt);
-
-        // Allows overriding thresholds based on the parent's thresholds.
-        SubscribeLocalEvent<TemperatureComponent, EntParentChangedMessage>(OnParentChange);
-        SubscribeLocalEvent<ContainerTemperatureDamageThresholdsComponent, ComponentStartup>(
-            OnParentThresholdStartup);
-        SubscribeLocalEvent<ContainerTemperatureDamageThresholdsComponent, ComponentShutdown>(
-            OnParentThresholdShutdown);
     }
 
     private void OnAtmosExposedUpdate(EntityUid uid, TemperatureComponent temperature,
@@ -115,117 +107,5 @@ public sealed class TemperatureSystem : SharedTemperatureSystem
         RaiseLocalEvent(uid, ref ev);
 
         args.Args.TemperatureDelta *= ev.Coefficient;
-    }
-
-    private void OnParentChange(EntityUid uid, TemperatureComponent component,
-        ref EntParentChangedMessage args)
-    {
-        var temperatureQuery = GetEntityQuery<TemperatureComponent>();
-        var transformQuery = GetEntityQuery<TransformComponent>();
-        var thresholdsQuery = GetEntityQuery<ContainerTemperatureDamageThresholdsComponent>();
-        // We only need to update thresholds if the thresholds changed for the entity's ancestors.
-        var oldThresholds = args.OldParent != null
-            ? RecalculateParentThresholds(args.OldParent.Value, transformQuery, thresholdsQuery)
-            : (null, null);
-        var newThresholds = RecalculateParentThresholds(transformQuery.GetComponent(uid).ParentUid, transformQuery, thresholdsQuery);
-
-        if (oldThresholds != newThresholds)
-        {
-            RecursiveThresholdUpdate(uid, temperatureQuery, transformQuery, thresholdsQuery);
-        }
-    }
-
-    private void OnParentThresholdStartup(EntityUid uid, ContainerTemperatureDamageThresholdsComponent component,
-        ComponentStartup args)
-    {
-        RecursiveThresholdUpdate(uid, GetEntityQuery<TemperatureComponent>(), GetEntityQuery<TransformComponent>(),
-            GetEntityQuery<ContainerTemperatureDamageThresholdsComponent>());
-    }
-
-    private void OnParentThresholdShutdown(EntityUid uid, ContainerTemperatureDamageThresholdsComponent component,
-        ComponentShutdown args)
-    {
-        RecursiveThresholdUpdate(uid, GetEntityQuery<TemperatureComponent>(), GetEntityQuery<TransformComponent>(),
-            GetEntityQuery<ContainerTemperatureDamageThresholdsComponent>());
-    }
-
-    /// <summary>
-    /// Recalculate and apply parent thresholds for the root entity and all its descendant.
-    /// </summary>
-    /// <param name="root"></param>
-    /// <param name="temperatureQuery"></param>
-    /// <param name="transformQuery"></param>
-    /// <param name="tempThresholdsQuery"></param>
-    private void RecursiveThresholdUpdate(EntityUid root, EntityQuery<TemperatureComponent> temperatureQuery,
-        EntityQuery<TransformComponent> transformQuery,
-        EntityQuery<ContainerTemperatureDamageThresholdsComponent> tempThresholdsQuery)
-    {
-        RecalculateAndApplyParentThresholds(root, temperatureQuery, transformQuery, tempThresholdsQuery);
-
-        var enumerator = Transform(root).ChildEnumerator;
-        while (enumerator.MoveNext(out var child))
-        {
-            RecursiveThresholdUpdate(child, temperatureQuery, transformQuery, tempThresholdsQuery);
-        }
-    }
-
-    /// <summary>
-    /// Recalculate parent thresholds and apply them on the uid temperature component.
-    /// </summary>
-    /// <param name="uid"></param>
-    /// <param name="temperatureQuery"></param>
-    /// <param name="transformQuery"></param>
-    /// <param name="tempThresholdsQuery"></param>
-    private void RecalculateAndApplyParentThresholds(EntityUid uid,
-        EntityQuery<TemperatureComponent> temperatureQuery, EntityQuery<TransformComponent> transformQuery,
-        EntityQuery<ContainerTemperatureDamageThresholdsComponent> tempThresholdsQuery)
-    {
-        if (!temperatureQuery.TryGetComponent(uid, out var temperature))
-        {
-            return;
-        }
-
-        var newThresholds = RecalculateParentThresholds(transformQuery.GetComponent(uid).ParentUid, transformQuery, tempThresholdsQuery);
-        temperature.ParentHeatDamageThreshold = newThresholds.Item1;
-        temperature.ParentColdDamageThreshold = newThresholds.Item2;
-    }
-
-    /// <summary>
-    /// Recalculate Parent Heat/Cold DamageThreshold by recursively checking each ancestor and fetching the
-    /// maximum HeatDamageThreshold and the minimum ColdDamageThreshold if any exists (aka the best value for each).
-    /// </summary>
-    /// <param name="initialParentUid"></param>
-    /// <param name="transformQuery"></param>
-    /// <param name="tempThresholdsQuery"></param>
-    private (float?, float?) RecalculateParentThresholds(
-        EntityUid initialParentUid,
-        EntityQuery<TransformComponent> transformQuery,
-        EntityQuery<ContainerTemperatureDamageThresholdsComponent> tempThresholdsQuery)
-    {
-        // Recursively check parents for the best threshold available
-        var parentUid = initialParentUid;
-        float? newHeatThreshold = null;
-        float? newColdThreshold = null;
-        while (parentUid.IsValid())
-        {
-            if (tempThresholdsQuery.TryGetComponent(parentUid, out var newThresholds))
-            {
-                if (newThresholds.HeatDamageThreshold != null)
-                {
-                    newHeatThreshold = Math.Max(newThresholds.HeatDamageThreshold.Value,
-                        newHeatThreshold ?? 0);
-                }
-
-                if (newThresholds.ColdDamageThreshold != null)
-                {
-                    newColdThreshold = Math.Min(newThresholds.ColdDamageThreshold.Value,
-                        newColdThreshold ?? float.MaxValue);
-                }
-            }
-
-            parentUid = transformQuery.GetComponent(parentUid).ParentUid;
-        }
-
-        return (newHeatThreshold, newColdThreshold);
     }
 }
